requirements.txt
flask==2.2.3
paho-mqtt==1.6.1